Prep for s19 (#2)
Updates the write-up for Spring, 2019.

The big changes are a restructuring of how things are turned in, and attempts to clarify some of the more confusing bits (e.g., in the discussion of combos).

I also tidied up and extended the use of footnotes.

I was confused about exactly what was going on with the combos example,
but trading some e-mails with @dolan-peter fixed that. Hopefully my changes
here will help clarify that for others.

I've re-structured the
"Checklist of what to do" in the hopes of making it clearer _where_
to turn things in as well as _what_ to turn in.